requirements.txt
# Core dependencies
fastapi==0.120.0
uvicorn==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.11

# Data processing
pandas==2.3.3
numpy==2.3.4

# Authentication
PyJWT==2.10.1
python-dotenv==1.1.1

# Dashboard
streamlit==1.44.0
plotly==6.3.1

# Development
pytest==8.4.2
requests==2.32.3
